{evaluation}: move expected runner tests into local_test
